Move UniTask init to the UnityEngine area
